python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nmnguyen130", email = "nguyenminhnguyen1324@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aioodbc>=0.5.0",
    "alembic>=1.18.4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "pyodbc>=5.3.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[tool.uv]
package = false
